+++ R/GenerateConceptSets.R
@@ -227,7 +227,7 @@ removeNonRelevantConcepts <- function(concepts, conditionPrompt, client, systemP
 #'
 #' @param phenotype              A text string denoting the condition of interest, e.g. 'Type I Diabetes Mellitus
 #'                               (T1DM)'. This string is used as input for the LLM.
-#' @param phenotypeDefinition Optionally prove a text blob with the definition and any other information about the
+#' @param clinicalDefinition Optionally prove a text blob with the definition and any other information about the
 #'                            phenotype.
 #' @param client                 An LLM client created using the `ellmer` package.
 #' @param vocabConnectionDetails Connection details for a database server hosting the OHDSI Vocabulary. Should be
@@ -247,14 +247,14 @@ removeNonRelevantConcepts <- function(concepts, conditionPrompt, client, systemP
 #' @export
 generateKeeperConceptSets <- function(
   phenotype,
-  phenotypeDefinition = NULL,
+  clinicalDefinition = NULL,
   client,
   vocabConnectionDetails,
   vocabDatabaseSchema
 ) {
   errorMessages <- checkmate::makeAssertCollection()
   checkmate::assertCharacter(phenotype, min.chars = 1, len = 1, add = errorMessages)
-  checkmate::assertCharacter(phenotypeDefinition, len = 1, null.ok = TRUE, add = errorMessages)
+  checkmate::assertCharacter(clinicalDefinition, len = 1, null.ok = TRUE, add = errorMessages)
   checkmate::assertR6(client, "Chat", add = errorMessages)
   checkmate::assertClass(vocabConnectionDetails, "ConnectionDetails", add = errorMessages)
   checkmate::assertCharacter(vocabDatabaseSchema, min.chars = 1, add = errorMessages)
@@ -276,7 +276,7 @@ generateKeeperConceptSets <- function(
     message(sprintf("Generating concept set %s", promptSet$name))
     conceptSet <- generateConceptSet(
       phenotype = phenotype,
-      phenotypeDefinition = phenotypeDefinition,
+      clinicalDefinition = clinicalDefinition,
       promptSet = promptSet,
       connection = connection,
       vocabDatabaseSchema = vocabDatabaseSchema,
@@ -299,7 +299,7 @@ generateKeeperConceptSets <- function(
 
       conceptSet <- generateConceptSet(
         phenotype = paste0("\n- ", paste(alternativeDiagnoses, collapse = "\n- ")),
-        phenotypeDefinition = NULL,
+        clinicalDefinition = NULL,
         promptSet = promptSet,
         connection = connection,
         vocabDatabaseSchema = vocabDatabaseSchema,
@@ -329,7 +329,7 @@ generateKeeperConceptSets <- function(
 
 
 generateConceptSet <- function(phenotype,
-                               phenotypeDefinition,
+                               clinicalDefinition,
                                promptSet,
                                client,
                                connection,
@@ -340,9 +340,9 @@ generateConceptSet <- function(phenotype,
 
   conditionPrompt <- sprintf("Condition: %s", phenotype)
   
-  if (!is.null(phenotypeDefinition)) {
+  if (!is.null(clinicalDefinition)) {
     conditionPrompt <- paste(conditionPrompt,
-                             sprintf("Definition: %s", phenotypeDefinition), sep = "\n\n")
+                             sprintf("Definition: %s", clinicalDefinition), sep = "\n\n")
   }
 
   message("- Generating initial term list using LLM")