+++ R/LlmLoop.R
@@ -20,7 +20,7 @@
 #' @param settings    Prompt creating settings as created using the [createPromptSettings] function.
 #' @param phenotypeName The name of the disease to use in the prompt. If not provided, the name in the Keeper input will
 #'                      be used.
-#' @param phenotypeDefinition Optionally prove a text blob with the definition and any other information about the
+#' @param clinicalDefinition Optionally prove a text blob with the definition and any other information about the
 #'                            phenotype.
 #' @param client      An LLM client created using the `ellmer` package.
 #' @param cacheFolder A folder where the LLM responses are cached. If the process terminates for some
@@ -46,7 +46,7 @@
 reviewCases <- function(keeper,
                         settings = createPromptSettings(),
                         phenotypeName = NULL,
-                        phenotypeDefinition = NULL,
+                        clinicalDefinition = NULL,
                         client,
                         cacheFolder) {
   errorMessages <- checkmate::makeAssertCollection()
@@ -63,16 +63,16 @@ reviewCases <- function(keeper,
   ), add = errorMessages)
   checkmate::assertClass(settings, "PromptSettings", add = errorMessages)
   checkmate::assertCharacter(phenotypeName, len = 1, null.ok = TRUE, add = errorMessages)
-  checkmate::assertCharacter(phenotypeDefinition, null.ok = TRUE, add = errorMessages)
+  checkmate::assertCharacter(clinicalDefinition, null.ok = TRUE, add = errorMessages)
   checkmate::assertR6(client, "Chat", add = errorMessages)
   checkmate::assertCharacter(cacheFolder, add = errorMessages)
   checkmate::reportAssertions(collection = errorMessages)
   
-  if (!is.null(phenotypeDefinition)) {
+  if (!is.null(clinicalDefinition)) {
     if (settings$legacy) {
-      warning("The phenotypeDefinition argument is ignored when using the legacy prompt settings.")
+      warning("The clinicalDefinition argument is ignored when using the legacy prompt settings.")
     }
-    phenotypeDefinition <- paste(phenotypeDefinition, collapse = "\n")
+    clinicalDefinition <- paste(clinicalDefinition, collapse = "\n")
   }
   
   startTime <- Sys.time()
@@ -116,7 +116,7 @@ reviewCases <- function(keeper,
     } else {
       systemPrompt <- createSystemPrompt(settings = settings, 
                                          phenotypeName = phenotype,
-                                         phenotypeDefinition = phenotypeDefinition)
+                                         clinicalDefinition = clinicalDefinition)
       if (settings$legacy) {
         prompt <- createLegacyPrompt(
           settings = settings,